use profileService for profile update & fix reload profile info gone issue
